// src/app/types/Match.ts
export type Match = {
    gameId: string;
    matchId: string;
    teamId: string;
    createdAt: number;
    updatedAt: number;
    score: string;
    teamWin: string;
    firstHalfScore: string;
    secondHalfScore: string;
    team: string;
    finalScore: string;
    teamHeadshots: string;
    teamKillDeathRatio: string;
    teamKillRoundRatio: string;
    teamHeadshotsPercentage: string;
    region: string;
    map: string;
    rounds: string;
    status: string;
    gameMode: string;
    premade: boolean;
    competitionId: string;
    date: number;
    played: string;
    game: string;
    winner: string;
    matchRound: string;
  };